move crlf to kubectl/util
