+++ pkg/kubectl/cmd/util/editor/BUILD
@@ -21,8 +21,8 @@ go_library(
         "//pkg/kubectl:go_default_library",
         "//pkg/kubectl/cmd/util:go_default_library",
         "//pkg/kubectl/resource:go_default_library",
+        "//pkg/kubectl/util/crlf:go_default_library",
         "//pkg/printers:go_default_library",
-        "//pkg/util/crlf:go_default_library",
         "//pkg/util/term:go_default_library",
         "//vendor/github.com/evanphx/json-patch:go_default_library",
         "//vendor/github.com/golang/glog:go_default_library",
